# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My Express Router API",
    "description": "API Docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "web-services-ok3j.onrender.com",
      "description": "Production Server"
    },
    {
      "url": "invidividual-project.onrender.com",
      "description": "individual Server"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/subscription/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "providerName": {
                  "example": "any"
                },
                "monthlyCost": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "renewalDate": {
                  "example": "any"
                },
                "paymentMethod": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "googleId": {
                  "example": "any"
                },
                "accountCreated": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}